restuarent_app/requirements.txt
altgraph==0.17.3
anaconda==0.0.1.1
anyio==4.2.0
appdirs==1.4.4
arabic-reshaper==3.0.0
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
auto-py-to-exe==2.37.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.2.0
comtypes==1.2.0
cryptography==41.0.2
cssselect2==0.7.0
Cython==3.0.11
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.3
django-extensions==4.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
docker==7.1.0
easy-colors==0.0.1
Eel==0.16.0
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.15.4
fqdn==1.5.1
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
googletrans==3.0.0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hl7apy==1.3.5
hpack==3.0.0
hstspreload==2023.1.1
html2pdf==0.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.5.2
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
openpyxl==3.1.5
optional-django==0.3.0
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
Pillow==10.0.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarmor==9.1.6
pyarmor.cli.core==7.6.6
pycparser==2.21
Pygments==2.17.2
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pyinstaller==5.13.1
pyinstaller-hooks-contrib==2023.7
pyparsing==3.1.1
pypdf==3.12.2
pypiwin32==223
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-escpos==3.1
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
react==4.3.0
referencing==0.32.0
reportlab==3.6.13
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.15.2
send-mail==0.0.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
uritools==4.0.1
urllib3==2.0.3
virtualenv==20.26.3
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
whichcraft==0.6.1
widgetsnbextension==4.0.9
xhtml2pdf==0.2.11
zope.event==5.0
zope.interface==6.0
